requirements.txt
numpy
matplotlib
opencv-python
pillow
scikit-learn
scipy
trimesh
plyfile
laspy
open3d
pyproj
pyquaternion
tqdm